Split ambiguous names into individual rules
