+++ src/rules/pycodestyle/helpers.rs
@@ -1,0 +1,3 @@
+pub fn is_ambiguous_name(name: &str) -> bool {
+    name == "l" || name == "I" || name == "O"
+}
