+++ src/rules/pycodestyle/rules/ambiguous_class_name.rs
@@ -1,0 +1,19 @@
+use crate::ast::types::Range;
+use crate::registry::Diagnostic;
+use crate::rules::pycodestyle::helpers::is_ambiguous_name;
+use crate::violations;
+
+/// E742
+pub fn ambiguous_class_name<F>(name: &str, locate: F) -> Option<Diagnostic>
+where
+    F: FnOnce() -> Range,
+{
+    if is_ambiguous_name(name) {
+        Some(Diagnostic::new(
+            violations::AmbiguousClassName(name.to_string()),
+            locate(),
+        ))
+    } else {
+        None
+    }
+}
